fix: FactsDB layers, manage CLI bindings, PR #997 review feedback
- Split FactsDB across facts-db-layer1/2/3 with thin facts-db-core re-export; add
  extracted modules (clusters, contradictions, housekeeping, etc.).
- Restore ManageBindings destructuring and imports (capturePluginError, fs/path,
  getEnv, cron helpers) so manage modules typecheck again.
- PR feedback: validate vectordb-optimize --older-than-days and list --limit;
  normalize errors in vector-count catch; tier breakdown includes structural;
  hybrid search applies category/entity/key/source/tier to merged results;
  cron backup log path shell-safe; use estimateStoredTokens in stats --efficiency.

Made-with: Cursor
