+++ extensions/memory-hybrid/backends/facts-db/schema-bootstrap.ts
@@ -1,0 +1,55 @@
+/**
+ * Minimal facts + FTS5 schema before migrations run (Issue #954 split).
+ */
+import type { DatabaseSync } from "node:sqlite";
+
+export function bootstrapFactsCoreSchema(db: DatabaseSync): void {
+  db.exec(`
+      CREATE TABLE IF NOT EXISTS facts (
+        id TEXT PRIMARY KEY,
+        text TEXT NOT NULL,
+        why TEXT,
+        category TEXT NOT NULL DEFAULT 'other',
+        importance REAL NOT NULL DEFAULT 0.5,
+        entity TEXT,
+        key TEXT,
+        value TEXT,
+        source TEXT NOT NULL DEFAULT 'conversation',
+        created_at INTEGER NOT NULL
+      )
+    `);
+
+  db.exec(`
+      CREATE VIRTUAL TABLE IF NOT EXISTS facts_fts USING fts5(
+        text,
+        category,
+        entity,
+        key,
+        value,
+        tokenize='porter unicode61'
+      )
+    `);
+
+  db.exec(`
+      CREATE TRIGGER IF NOT EXISTS facts_ai AFTER INSERT ON facts BEGIN
+        INSERT INTO facts_fts(rowid, text, category, entity, key, value)
+        VALUES (new.rowid, new.text, new.category, new.entity, new.key, new.value);
+      END;
+
+      CREATE TRIGGER IF NOT EXISTS facts_ad AFTER DELETE ON facts BEGIN
+        DELETE FROM facts_fts WHERE rowid = old.rowid;
+      END;
+
+      CREATE TRIGGER IF NOT EXISTS facts_au AFTER UPDATE ON facts BEGIN
+        DELETE FROM facts_fts WHERE rowid = old.rowid;
+        INSERT INTO facts_fts(rowid, text, category, entity, key, value)
+        VALUES (new.rowid, new.text, new.category, new.entity, new.key, new.value);
+      END
+    `);
+
+  db.exec(`
+      CREATE INDEX IF NOT EXISTS idx_facts_category ON facts(category);
+      CREATE INDEX IF NOT EXISTS idx_facts_entity ON facts(entity);
+      CREATE INDEX IF NOT EXISTS idx_facts_created ON facts(created_at);
+    `);
+}
